requirements.txt
flake8==3.9.2
pytest==6.2.5
twine==3.4.2
build==0.5.1
wheel==0.37.0
requests==2.26.0
pytest-mock==3.6.1
PyJWT==2.2.0
pyinstaller==4.5.1